# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tonatiuh Trading SaaS API
  version: 1.0.0
  description: Multi-tenant API for accounts, subscriptions, exchange connections, and trading bots.
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []
...
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      operationId: register
      security: []
      requestBody: {$ref: '#/components/requestBodies/Register'}
      responses:
        '201': {description: Account created; email verification required}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: login
      security: []
      requestBody: {$ref: '#/components/requestBodies/Login'}
      responses:
        '200': {$ref: '#/components/responses/Session'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshSession
      security: []
      requestBody: {$ref: '#/components/requestBodies/RefreshToken'}
      responses:
        '200': {$ref: '#/components/responses/Session'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      security: []
      requestBody: {$ref: '#/components/requestBodies/RefreshToken'}
      responses:
        '204': {description: Session revoked}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/verify-email:
    post:
      tags: [Auth]
      operationId: verifyEmail
      security: []
      requestBody: {$ref: '#/components/requestBodies/AccountToken'}
      responses:
        '200': {description: Email verified}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/resend-verification:
    post:
      tags: [Auth]
      operationId: resendVerification
      security: []
      requestBody: {$ref: '#/components/requestBodies/Email'}
      responses:
        '202': {description: Request accepted}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      operationId: forgotPassword
      security: []
      requestBody: {$ref: '#/components/requestBodies/Email'}
      responses:
        '202': {description: Request accepted}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token, password]
              properties:
                token: {type: string}
                password: {type: string, minLength: 12, maxLength: 128, format: password}
      responses:
        '200': {description: Password replaced and sessions revoked}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/cancel-deletion:
    post:
      tags: [Auth]
      operationId: cancelAccountDeletion
      security: []
      requestBody: {$ref: '#/components/requestBodies/Login'}
      responses:
        '200': {description: Scheduled account deletion cancelled}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/me:
    get:
      tags: [Auth]
      operationId: currentAccount
      responses:
        '200': {description: Current account and organization}
        default: {$ref: '#/components/responses/Error'}
    delete:
      tags: [Auth]
      operationId: scheduleAccountDeletion
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, additionalProperties: false, required: [password], properties: {password: {type: string, format: password}}}}}
      responses:
        '202': {description: Account deletion scheduled after the recovery period}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/me/export:
    get:
      tags: [Auth]
      operationId: exportAccountData
      responses:
        '200': {description: Portable account data export}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/auth/me/data-requests:
    get:
      tags: [Compliance]
      operationId: listMyDataSubjectRequests
      responses: {'200': {description: Current user data-subject requests}, default: {$ref: '#/components/responses/Error'}}
    post:
      tags: [Compliance]
      operationId: createDataSubjectRequest
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, additionalProperties: false, required: [kind, details], properties: {kind: {type: string, enum: [ACCESS, RECTIFY, RESTRICT, OBJECT]}, details: {type: string, minLength: 1, maxLength: 2000}}}}}
      responses: {'202': {description: Data-subject request accepted with a response deadline}, default: {$ref: '#/components/responses/Error'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    metricsToken: {type: http, scheme: bearer, bearerFormat: opaque metrics token}
    emailWebhookToken: {type: http, scheme: bearer, bearerFormat: opaque provider webhook token}
...
  responses:
    Error:
      description: Standard API error
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
    Session:
      description: Access and rotating refresh tokens
      content: {application/json: {schema: {$ref: '#/components/schemas/Session'}}}
  requestBodies:
    Register:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/Register'}}}
    Login:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/Login'}}}
    RefreshToken:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, required: [refreshToken], properties: {refreshToken: {type: string}}}}}
    AccountToken:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, required: [token], properties: {token: {type: string}}}}}
    Email:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, required: [email], properties: {email: {type: string, format: email, maxLength: 320}}}}}
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties: {code: {type: string}, message: {type: string}, requestId: {type: string}, details: {}}
    Session:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties: {accessToken: {type: string, writeOnly: true}, refreshToken: {type: string, writeOnly: true}, expiresIn: {type: integer}}
...
    Register:
      type: object
      additionalProperties: false
      required: [email, password, displayName, acceptTerms, acceptPrivacy]
      properties: {email: {type: string, format: email, maxLength: 320}, password: {type: string, format: password, minLength: 12, maxLength: 128}, displayName: {type: string, minLength: 2, maxLength: 120}, organizationName: {type: string, maxLength: 120}, acceptTerms: {type: boolean}, acceptPrivacy: {type: boolean}, termsVersion: {type: string, minLength: 1, maxLength: 100}, privacyVersion: {type: string, minLength: 1, maxLength: 100}}
    Login:
      type: object
      additionalProperties: false
      required: [email, password]
      properties: {email: {type: string, format: email}, password: {type: string, format: password}}
...